feat: localize settings, ride history, wallet, stops, ratings, and cleanup unused localization keys

+++ lib/feature/home/domain/entities/ride_request_entity.dart
@@ -21,7 +21,7 @@ class RideRequestEntity with _$RideRequestEntity {
     required List<RideRequestStopEntity> stops,
     int? rideRequestId,
     double? distanceKm,
-    int? durationMin,
+    double? durationMin,
     bool? isNight,
     String? estimatedFare,
     String? currency,
